Move REQUIRED and OPTIONAL constants to ._common.
