+++ dataclass_builder/factory.py
@@ -101,8 +101,8 @@ or `OPTIONAL` constant
 from typing import (Any, Callable, Dict, Mapping, MutableMapping, Optional,
                     Sequence, TYPE_CHECKING, cast)
 
-from ._common import (_settable_fields, _required_fields, _optional_fields,
-                      _is_required)
+from ._common import (REQUIRED, OPTIONAL, _is_required,
+                      _settable_fields, _required_fields, _optional_fields)
 from .exceptions import UndefinedFieldError, MissingFieldError
 
 if TYPE_CHECKING:
@@ -110,7 +110,7 @@ if TYPE_CHECKING:
 else:
     from dataclasses import is_dataclass
 
-__all__ = ['dataclass_builder', 'REQUIRED', 'OPTIONAL']
+__all__ = ['dataclass_builder']
 
 
 class _MissingType:
@@ -118,34 +118,20 @@ class _MissingType:
         return 'MISSING'
 
 
-class _RequiredType:
-    def __repr__(self) -> str:
-        return 'REQUIRED'
-
-
-class _OptionalType:
-    def __repr__(self) -> str:
-        return 'OPTIONAL'
-
-
-MISSING = _MissingType()
-
-REQUIRED = _RequiredType()
-
-OPTIONAL = _OptionalType()
+_MISSING = _MissingType()
 
 
 # copied (and modified) from dataclasses._create_fn to avoid dependency on
 # private functions in dataclasses
 def _create_fn(name: str, args: Sequence[str], body: Sequence[str],
                env: Optional[Dict[str, Any]] = None, *,
-               return_type: Any = MISSING) \
+               return_type: Any = _MISSING) \
         -> Callable[..., Any]:
     locals_: MutableMapping[str, Any] = {}
     return_annotation = ''
     if env is None:
         env = {}
-    if return_type is not MISSING:
+    if return_type is not _MISSING:
         env['_return_type'] = return_type
         return_annotation = '->_return_type'
     args = ', '.join(args)
