+++ tests/test_common.py
@@ -1,9 +1,19 @@
 from dataclasses import fields
 
 from dataclass_builder._common import (
+    REQUIRED, OPTIONAL,
     _is_settable, _is_required, _is_optional,
     _settable_fields, _required_fields, _optional_fields)
 from tests.conftest import PixelCoord, Point, Circle, Types  # type: ignore
+
+
+def test_constants():
+    assert REQUIRED == REQUIRED
+    assert REQUIRED != OPTIONAL
+    assert OPTIONAL == OPTIONAL
+    assert OPTIONAL != REQUIRED
+    assert repr(REQUIRED) == 'REQUIRED'
+    assert repr(OPTIONAL) == 'OPTIONAL'
 
 
 def test_is_settable():
