+++ tests/test_factory.py
@@ -2,7 +2,7 @@ import pytest  # type: ignore
 from typing import Any, get_type_hints
 import dataclasses
 
-from dataclass_builder.factory import _create_fn, MISSING
+from dataclass_builder.factory import _create_fn, _MISSING
 from dataclass_builder import (MissingFieldError, UndefinedFieldError,
                                REQUIRED, OPTIONAL,
                                dataclass_builder, build, fields)
@@ -12,15 +12,9 @@ from tests.conftest import (Circle, PixelCoord, Point,  # type: ignore
 
 
 def test_constants():
-    assert MISSING == MISSING
-    assert MISSING != REQUIRED
-    assert REQUIRED == REQUIRED
-    assert REQUIRED != OPTIONAL
-    assert OPTIONAL == OPTIONAL
-    assert OPTIONAL != MISSING
-    assert repr(MISSING) == 'MISSING'
-    assert repr(REQUIRED) == 'REQUIRED'
-    assert repr(OPTIONAL) == 'OPTIONAL'
+    assert _MISSING == _MISSING
+    assert _MISSING != 1
+    assert repr(_MISSING) == 'MISSING'
 
 
 def test_create_fn():
